requirements/base.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
uvicorn[standard]
gunicorn
